requirements-vercel.txt
Django==6.0.1
python-dotenv==1.0.0
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
Pillow
gunicorn
whitenoise[brotli]
reportlab
openpyxl
python-docx
psycopg2-binary
dj-database-url
pypdf
